+++ dashboard/pages/1_About_this_site.py
@@ -1,15 +1,45 @@
 """Create and host a streamlit dashboard"""
 
-from os import environ as ENV
-
-import altair as alt
-import pandas as pd
-from psycopg2 import connect
-from psycopg2.extras import RealDictCursor
+import pydeck as pdk
 from dotenv import load_dotenv
+from psycopg2.extras import RealDictCursor
+from psycopg2 import connect
+import pandas as pd
+import altair as alt
+from os import environ as ENV
 import streamlit as st
 from vega_datasets import data
-import pydeck as pdk
+
+
+def intro():
+
+    st.write("# Welcome to the Climate Monitor! :sun_small_cloud:")
+    st.markdown(
+        f"""
+        **👈 Select a page from the menu on the left** to explore the dashboard!
+
+        **Home**: Quick search - get the coming forecast and find weather alerts by location!
+
+        **Explore**: Explore the current climate across the UK.
+
+        **Alerts across the UK**: Get the latest weather alerts, including floods and air quality warnings.
+
+        ### Want to sign up to daily newsletters, or receive notifications about weather alerts near you?
+
+        - Then check out the **Sign Up** page!""")
+    st.page_link('pages/4_Sign_Up.py')
+    st.markdown("""
+
+        ### Contributors
+        Project manager: **Dom Chambers**
+
+        Quality Assurance: **Arjun Babhania**
+
+        Architect: **Nathan McKittrick**
+
+        Architect: **Dana Weetman**
+    """
+                )
 
 
 @st.cache_resource
@@ -45,6 +75,9 @@ def get_location_data(_conn) -> pd.DataFrame:
                     ON (l.county_id=c.county_id)
                     JOIN country as co
                     ON (c.country_id=co.country_id)
+                    WHERE l.loc_id IN (
+                    SELECT loc_id from weather_report
+                    )
                     """)
 
         rows = cur.fetchall()
@@ -93,7 +126,7 @@ def uk_map(loc_data, tooltips):
         type='mercator',
         scale=1500,                          # Magnify
         center=[-2, 54],                     # [lon, lat]
-        clipExtent=[[0, 0], [500, 500]],    # [[left, top], [right, bottom]]
+        clipExtent=[[0, 0], [600, 600]],    # [[left, top], [right, bottom]]
     ).properties(
         width=500,
         height=500
@@ -116,15 +149,18 @@ def uk_map(loc_data, tooltips):
 
 
 if __name__ == "__main__":
-
+    intro()
     load_dotenv()
     with connect_to_db(dict(ENV)) as connection:
         forecast_data = get_data_from_db(connection, 'forecast')
         forecast_d = get_location_forecast_data(connection)
         st.title('About the Data')
+        locations = get_location_data(connection)
+        print(locations)
         w_map = uk_map(get_location_data(
             connection), ['location', 'county'])
         st.altair_chart(w_map, use_container_width=True)
-
-        w_map = uk_map(forecast_d, ['location', 'weather'])
-        st.altair_chart(w_map, use_container_width=True)
+        st.write(
+            f"This climate monitor is currently tracking {locations['location'].count()} locations across the UK.")
+        st.write(
+            "If you want to add more locations, 👈 navigate to the **Sign Up** page!")
